Move graphics handling into seperate header files. This just makes it easier to contribute to the project without having a graphicl calculator or eumlator.
